Add question to DB with survey_id swell
Not the most efficient way possible I don’t think.. To do this I added
an new page surveys/add, which links to the newly created ‘add’
function in the survey controller. To add questions, the user does this
from the index of their survey. The aim now that this is done is to
leave it as it is for now as it ‘works’ and then later, maybe using
javascript have a way to dynamically change question type, using a
dropdown that loads different types. Not sure how this would work. For
now, I shall look at editing questions and listing the questions.